style renamed to readability
